api/pvt/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
dnspython==2.1.0
pymongo==3.11.3
python-dotenv==0.17.0